requirements.txt
# Updated requirements for VideoAnomalyCLIP
torch==2.0.0
torchvision==0.15.1
numpy>=1.21.0
scipy==1.9.1
scikit-learn==1.2.2
scikit-image==0.20.0

# Video processing
decord>=0.6.0
opencv-python>=4.5.0

# Configuration and utilities
pyyaml>=6.0
tqdm==4.65.0
tabulate

# Image processing and visualization
Pillow>=9.0.0
matplotlib>=3.5.0
seaborn==0.12.2

# Deep learning utilities
timm==0.6.13
torchsummary==1.5.1

# For CLIP model
ftfy
regex